后台管理啊，好累啊，快点结束吧

+++ src/main/java/com/sjzxy/fwpt/service/impl/ReportServiceImpl.java
@@ -25,7 +25,7 @@ public class ReportServiceImpl implements ReportService {
 
     @Override
     public void delQuesReport(int id) {
-        reportRepository.deleteById((long) id);
+        reportRepository.deleteById(id);
     }
 
     @Override
